Improve ProposalList loading (#192)
