+++ apps/app/src/lib/i18n/dictionaries/es.json
@@ -107,7 +107,6 @@
   "Participate": "Participar",
   "Load More": "Cargar más",
   "Loading...": "Cargando...",
-  "My proposals •": "Mis propuestas •",
   "My proposals": "Mis propuestas",
   "All proposals": "Todas las propuestas",
   "All categories": "Todas las categorías",
@@ -235,6 +234,6 @@
   "Shortlisted": "Preseleccionada",
   "Reject from shortlist": "Rechazar de la preselección",
   "Shortlist for voting": "Preseleccionar para votación",
-  "Shortlisted proposals •": "Propuestas preseleccionadas •",
+  "Shortlisted proposals": "Propuestas preseleccionadas",
   "Are you sure you want to delete this proposal? This action cannot be undone.": "¿Estás seguro de que quieres eliminar esta propuesta? Esta acción no se puede deshacer."
 }
