+++ apps/app/src/lib/i18n/dictionaries/fr.json
@@ -108,7 +108,6 @@
   "Participate": "Participer",
   "Load More": "Montrer plus",
   "Loading...": "En train de charger…",
-  "My proposals •": "Mes propositions •",
   "My proposals": "Mes propositions",
   "All proposals": "Toutes les propositions",
   "All categories": "Toutes les catégories",
@@ -236,6 +235,6 @@
   "Shortlisted": "Présélectionné",
   "Reject from shortlist": "Rejeter de la présélection",
   "Shortlist for voting": "Présélectionner pour le vote",
-  "Shortlisted proposals •": "Propositions présélectionnées •",
+  "Shortlisted proposals": "Propositions présélectionnées",
   "Are you sure you want to delete this proposal? This action cannot be undone.": "Êtes-vous sûr de vouloir supprimer cette proposition ? Cette action ne peut pas être annulée."
 }
