+++ apps/app/src/lib/i18n/dictionaries/pt.json
@@ -108,7 +108,6 @@
   "Participate": "Participar",
   "Load More": "Carregar mais",
   "Loading...": "Carregando...",
-  "My proposals •": "Minhas propostas •",
   "My proposals": "Minhas propostas",
   "All proposals": "Todas as propostas",
   "All categories": "Todas as categorias",
@@ -236,6 +235,6 @@
   "Shortlisted": "Pré-selecionada",
   "Reject from shortlist": "Rejeitar da pré-seleção",
   "Shortlist for voting": "Pré-selecionar para votação",
-  "Shortlisted proposals •": "Propostas pré-selecionadas •",
+  "Shortlisted proposals": "Propostas pré-selecionadas",
   "Are you sure you want to delete this proposal? This action cannot be undone.": "Tem certeza de que deseja excluir esta proposta? Esta ação não pode ser desfeita."
 }
